+++ src/openassetio-python/cmodule/src/managerApi/ManagerInterfaceBinding.cpp
@@ -119,6 +119,16 @@ struct PyManagerInterface : ManagerInterface {
                                   errorCallback);
   }
 
+  void entityTraits(const EntityReferences& entityReferences,
+                    const access::EntityTraitsAccess entityTraitsAccess,
+                    const ContextConstPtr& context, const HostSessionPtr& hostSession,
+                    const EntityTraitsSuccessCallback& successCallback,
+                    const BatchElementErrorCallback& errorCallback) override {
+    OPENASSETIO_PYBIND11_OVERRIDE(void, ManagerInterface, entityTraits, entityReferences,
+                                  entityTraitsAccess, context, hostSession, successCallback,
+                                  errorCallback);
+  }
+
   void defaultEntityReference(const trait::TraitSets& traitSets,
                               const access::DefaultEntityAccess defaultEntityAccess,
                               const ContextConstPtr& context, const HostSessionPtr& hostSession,
@@ -210,7 +220,8 @@ void registerManagerInterface(const py::module& mod) {
       .value("kPublishing", ManagerInterface::Capability::kPublishing)
       .value("kRelationshipQueries", ManagerInterface::Capability::kRelationshipQueries)
       .value("kExistenceQueries", ManagerInterface::Capability::kExistenceQueries)
-      .value("kDefaultEntityReferences", ManagerInterface::Capability::kDefaultEntityReferences);
+      .value("kDefaultEntityReferences", ManagerInterface::Capability::kDefaultEntityReferences)
+      .value("kEntityTraitIntrospection", ManagerInterface::Capability::kEntityTraitIntrospection);
 
   pyManagerInterface.attr("kCapabilityNames") = ManagerInterface::kCapabilityNames;
 
@@ -246,6 +257,10 @@ void registerManagerInterface(const py::module& mod) {
            py::arg("context").none(false), py::arg("hostSession").none(false),
            py::arg("successCallback"), py::arg("errorCallback"),
            py::call_guard<py::gil_scoped_release>{})
+      .def("entityTraits", &ManagerInterface::entityTraits, py::arg("entityReferences"),
+           py::arg("entityTraitsAccess"), py::arg("context").none(false),
+           py::arg("hostSession").none(false), py::arg("successCallback"),
+           py::arg("errorCallback"), py::call_guard<py::gil_scoped_release>{})
       .def("hasCapability", &ManagerInterface::hasCapability, py::arg("capability"),
            py::call_guard<py::gil_scoped_release>{})
       .def("updateTerminology", &ManagerInterface::updateTerminology, py::arg("terms"),